services/messaging/requirements.txt
Flask
Flask-SocketIO
eventlet
SQLAlchemy
psycopg2-binary
python-dotenv
flask-cors
kafka-python
python-dotenv